Refactor and move things a bit
